+++ lib/src/sync/optimistic.rs
@@ -159,6 +159,7 @@ struct OptimisticSyncInner<TRq, TSrc, TBl> {
         verification_queue::VerificationQueue<(RequestId, TRq), RequestSuccessBlock<TBl>>,
 
     /// Justifications, if any, of the block that has just been verified.
+    // TODO: clean up when a source is removed
     pending_encoded_justifications: vec::IntoIter<([u8; 4], Vec<u8>, SourceId)>,
 
     /// Identifier to assign to the next request.
@@ -209,11 +210,6 @@ struct Source<TSrc> {
 
     /// Best block that the source has reported having.
     best_block_number: u64,
-
-    /// If `true`, this source is banned and shouldn't use be used to request blocks.
-    /// Note that the ban is lifted if the source is removed. This ban isn't meant to be a line of
-    /// defense against malicious peers but rather an optimization.
-    banned: bool,
 
     /// Number of requests that use this source.
     num_ongoing_requests: u32,
@@ -383,7 +379,6 @@ impl<TRq, TSrc, TBl> OptimisticSync<TRq, TSrc, TBl> {
             Source {
                 user_data: source,
                 best_block_number,
-                banned: false,
                 num_ongoing_requests: 0,
             },
         );
@@ -657,16 +652,26 @@ impl<TRq, TSrc, TBl> OptimisticSync<TRq, TSrc, TBl> {
             .unwrap()
             .num_ongoing_requests -= 1;
 
-        self.inner.sources.get_mut(&source_id).unwrap().banned = true;
-
-        // If all sources are banned, unban them.
-        if self.inner.sources.iter().all(|(_, s)| s.banned) {
-            for src in self.inner.sources.values_mut() {
-                src.banned = false;
-            }
-        }
-
         user_data
+    }
+
+    /// Returns the [`SourceId`] that is expected to fulfill the given request.
+    ///
+    /// # Panic
+    ///
+    /// Panics if the [`RequestId`] is invalid.
+    ///
+    pub fn request_source_id(&self, request_id: RequestId) -> SourceId {
+        if let Some((src, _)) = self.inner.obsolete_requests.get(&request_id) {
+            *src
+        } else {
+            self.inner
+                .verification_queue
+                .requests()
+                .find(|(rq, _)| rq.0 == request_id)
+                .unwrap()
+                .1
+        }
     }
 
     /// Process the next block in the queue of verification.
@@ -850,24 +855,12 @@ impl<TRq, TSrc, TBl> BlockVerify<TRq, TSrc, TBl> {
                         scale_encoded_extrinsics: block.scale_encoded_extrinsics,
                         verified_header,
                         scale_encoded_justifications: block.scale_encoded_justifications,
-                        source_id,
                     },
                     new_best_hash,
                     new_best_number,
                 }
             }
             Err(reason) => {
-                if let Some(src) = self.inner.sources.get_mut(&source_id) {
-                    src.banned = true;
-                }
-
-                // If all sources are banned, unban them.
-                if self.inner.sources.iter().all(|(_, s)| s.banned) {
-                    for src in self.inner.sources.values_mut() {
-                        src.banned = false;
-                    }
-                }
-
                 self.inner.make_requests_obsolete(&self.chain);
 
                 let previous_best_height = self.chain.best_block_header().number;
@@ -924,7 +917,6 @@ pub struct BlockVerifySuccess<TRq, TSrc, TBl> {
     verified_header: blocks_tree::VerifiedHeader,
     scale_encoded_extrinsics: Vec<Vec<u8>>,
     scale_encoded_justifications: Vec<([u8; 4], Vec<u8>)>,
-    source_id: SourceId,
 }
 
 impl<TRq, TSrc, TBl> BlockVerifySuccess<TRq, TSrc, TBl> {
@@ -994,19 +986,7 @@ impl<TRq, TSrc, TBl> BlockVerifySuccess<TRq, TSrc, TBl> {
 
     /// Reject the block and mark it as bad.
     pub fn reject_bad_block(mut self) -> OptimisticSync<TRq, TSrc, TBl> {
-        if let Some(src) = self.parent.inner.sources.get_mut(&self.source_id) {
-            src.banned = true;
-        }
-
-        // If all sources are banned, unban them.
-        if self.parent.inner.sources.iter().all(|(_, s)| s.banned) {
-            for src in self.parent.inner.sources.values_mut() {
-                src.banned = false;
-            }
-        }
-
         self.parent.inner.make_requests_obsolete(&self.parent.chain);
-
         self.parent
     }
 
@@ -1041,6 +1021,17 @@ pub struct JustificationVerify<TRq, TSrc, TBl> {
 }
 
 impl<TRq, TSrc, TBl> JustificationVerify<TRq, TSrc, TBl> {
+    /// Returns the source the justification was obtained from.
+    pub fn sender(&self) -> (SourceId, &TSrc) {
+        let (_, _, source_id) = self
+            .inner
+            .pending_encoded_justifications
+            .as_slice()
+            .first()
+            .unwrap();
+        (*source_id, &self.inner.sources[source_id].user_data)
+    }
+
     /// Verify the justification.
     ///
     /// A randomness seed must be provided and will be used during the verification. Note that the
@@ -1052,7 +1043,7 @@ impl<TRq, TSrc, TBl> JustificationVerify<TRq, TSrc, TBl> {
         OptimisticSync<TRq, TSrc, TBl>,
         JustificationVerification<TBl>,
     ) {
-        let (consensus_engine_id, justification, source_id) =
+        let (consensus_engine_id, justification, _) =
             self.inner.pending_encoded_justifications.next().unwrap();
 
         let mut apply = match self.chain.verify_justification(
@@ -1062,17 +1053,6 @@ impl<TRq, TSrc, TBl> JustificationVerify<TRq, TSrc, TBl> {
         ) {
             Ok(a) => a,
             Err(error) => {
-                if let Some(source) = self.inner.sources.get_mut(&source_id) {
-                    source.banned = true;
-                }
-
-                // If all sources are banned, unban them.
-                if self.inner.sources.iter().all(|(_, s)| s.banned) {
-                    for src in self.inner.sources.values_mut() {
-                        src.banned = false;
-                    }
-                }
-
                 let chain = blocks_tree::NonFinalizedTree::new(
                     self.inner.finalized_chain_information.clone(),
                 );